// src/app/models/noticias.ts
export class Noticias {
    $key: string;
    tittle: string;
    phrase: string;
    autor: string;
    plantilla: string;
    date: string;
    part1: string;
    part2: string;
    part3: string;
    part4: string;
    part5: string;
    principalImage: string;
    image1: string;
    image2: string;
    image3: string;
    youtube: string;
}
export class NoticiasPlantilla6 {
    $key: string;
    tittle: string;
    phrase: string;
    autor: string;
    plantilla: string;
    date: string;
    part1: string;
    part2: string;
    part3: string;
    part4: string;
    part5: string;
    part6: string;
    part7: string;
    principalImage: string;
    image1: string;
    image2: string;
    image3: string;
    image4: string;
    image5: string;
    image6: string;
    youtube: string;
}